requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
pytest
pytest-cov
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator
pydantic[email]
python-multipart